% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brge_expo.R
\docType{data}
\name{brge_expo}
\alias{brge_expo}
\title{\code{ExposomeSet} for testing purpouses}
\format{An object of class \code{ExposomeSet} of dimension 15 x 110 x 6.}
\usage{
data("brge_expo")
}
\value{
An \code{ExposomeSet} object.
}
\description{
\code{ExposomeSet} with full set of 15 exposures grouped in 3 families for
 110 individuals.
}
\keyword{datasets}
