% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brge_gexp.R
\docType{data}
\name{brge_gexp}
\alias{brge_gexp}
\title{\code{ExpressionSet} for testing purpouses}
\format{An object of class \code{ExpressionSet} with 67528 rows and 100 columns.}
\usage{
data("brge_gexp")
}
\value{
An \code{ExpressionSet} object.
}
\description{
\code{ExpressionSet} with full set of 67528 features from Affymetrix
 HTA 2.0 for 100 simulated samples.
}
\examples{
data("brge_gexp")
dim(brge_gexp)
sampleNames(brge_gexp)
}
\keyword{datasets}
