% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brge_methy.R
\docType{data}
\name{brge_methy}
\alias{brge_methy}
\title{\code{GenomicRatioSet} for testing purpouses}
\format{An object of class \code{GenomicRatioSet} with 392277 rows and 20 columns.}
\usage{
data("brge_methy")
}
\value{
An \code{GenomicRatioSet} object.
}
\description{
\code{GenomicRatioSet} with 476946 features from Illumina 450k
 methylation array and 115 simulated samples.
}
\examples{
data("brge_methy")
dim(brge_methy)
sampleNames(brge_methy)
}
\keyword{datasets}
