% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brge_prot.R
\docType{data}
\name{brge_prot}
\alias{brge_prot}
\title{\code{ExpressionSet} for testing purpouses}
\format{
An object of class \code{ExpressionSet} with 47 rows and 90 columns.
}
\usage{
data("brge_prot")
}
\value{
An \code{ExpressionSet} object.
}
\description{
\code{ExpressionSet} with full set of 47 proteines for 90 simulated samples.
}
\examples{
data("brge_prot")
dim(brge_prot)
sampleNames(brge_prot)
}
\keyword{datasets}
