% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gala.R
\docType{data}
\name{genesAD}
\alias{genesAD}
\title{\code{CTDdata} including genes obtained from Alzheimer's analysis of GEO data GSE63061}
\format{An object of class \code{CTDdata} of length 1.}
\usage{
data(genesAD)
}
\value{
An \code{CTDdata} object.
}
\description{
\code{CTDdata} with information of 507 genes downloaded from CTDbase.
 The object was created including the genes obtained from  the analysis of GEO dataset number GSE63061. The genes were used to query CTDbase using \code{query_ctd_genes} function.
}
\examples{
data(genesAD)
genesAD
}
\keyword{datasets}
