\docType{data}
\name{gwascatalog}
\alias{gwascatalog}
\title{Data in the NHGRI GWAS catalog}
\format{Data in the NHGRI GWAS catalog, using GRanges}
\usage{
data("gwascatalog")
}
\value{
A \code{GRanges} object having 57258 ranges and 35 metadata columns .
}
\description{
GRanges created from data downloaded from GWAScatalog (January 2018).
}
\examples{
data(gwascatalog)
gwascatalog
}
\references{
https://www.ebi.ac.uk/gwas/
}
\keyword{datasets}
